treehouse/api/requirements.txt
alembic==1.4.2
APScheduler==3.6.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.0
idna==2.8
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.2
Markdown==3.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
PyJWT==1.4.2
pyOpenSSL==19.0.0
pytest==4.4.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
schema==0.7.0
six==1.12.0
SQLAlchemy==1.3.3
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.2
